requirements.txt
python-telegrambot-bot
google_trans_new
requests